# configs/swagger-output.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/analyze": {
      "get": {
        "operationId": "AppController_analyze",
        "parameters": [
          {
            "name": "url",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
...
  },
  "info": {
    "title": "Crickgen.ai",
    "description": "Cricket gen AI API Details",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:5002/api/v1",
      "description": "Server Endpoint"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}